Decentralize mocks
