+++ packages/framework/esm-config/jest.config.js
@@ -6,7 +6,7 @@ module.exports = {
   moduleNameMapper: {
     '@openmrs/esm-context': '<rootDir>/__mocks__/openmrs-esm-context.mock.tsx',
     '@openmrs/esm-globals': '<rootDir>/__mocks__/openmrs-esm-globals.mock.tsx',
-    '@openmrs/esm-state': '<rootDir>/__mocks__/openmrs-esm-state.mock.tsx',
+    '@openmrs/esm-state': '@openmrs/esm-state/mock',
   },
   testEnvironment: 'jsdom',
   testEnvironmentOptions: {
