+++ packages/framework/esm-react-utils/jest.config.js
@@ -6,8 +6,7 @@ module.exports = {
   moduleNameMapper: {
     '^lodash-es/(.*)$': 'lodash/$1',
     '@openmrs/esm-error-handling': '<rootDir>/__mocks__/openmrs-esm-error-handling.mock.ts',
-    '@openmrs/esm-state': '<rootDir>/__mocks__/openmrs-esm-state.mock.ts',
-    '@openmrs/esm-styleguide': '<rootDir>/__mocks__/openmrs-esm-styleguide.mock.tsx',
+    '@openmrs/esm-state': '@openmrs/esm-state/mock',
     dexie: require.resolve('dexie'),
   },
   testEnvironment: 'jsdom',
