+++ packages/framework/esm-react-utils/package.json
@@ -47,6 +47,7 @@
     "@openmrs/esm-config": "5.x",
     "@openmrs/esm-error-handling": "5.x",
     "@openmrs/esm-extensions": "5.x",
+    "@openmrs/esm-feature-flags": "5.x",
     "@openmrs/esm-globals": "5.x",
     "@openmrs/esm-navigation": "5.x",
     "dayjs": "1.x",
