+++ packages/framework/esm-styleguide/jest.config.js
@@ -7,7 +7,7 @@ module.exports = {
     '\\.(s?css)$': 'identity-obj-proxy',
     '^@carbon/icons-react/es/(.*)$': '@carbon/icons-react/lib/$1',
     '^@carbon/charts': 'identity-obj-proxy',
-    '@openmrs/esm-framework': '@openmrs/esm-framework/mock',
+    '@openmrs/esm-react-utils': '@openmrs/esm-react-utils/mock',
     '@openmrs/esm-translations': '@openmrs/esm-translations/mock',
     '^lodash-es/(.*)$': 'lodash/$1',
     dexie: require.resolve('dexie'),
