+++ packages/framework/esm-styleguide/src/patient-banner/contact-details/types.ts
@@ -1,4 +1,4 @@
-import { type OpenmrsResource } from '@openmrs/esm-framework';
+import { type OpenmrsResource } from '@openmrs/esm-api';
 
 export interface Location {
   uuid: string;
