+++ packages/framework/esm-styleguide/src/patient-banner/contact-details/useRelationships.ts
@@ -1,4 +1,4 @@
-import { openmrsFetch, restBaseUrl } from '@openmrs/esm-framework';
+import { openmrsFetch, restBaseUrl } from '@openmrs/esm-api';
 import useSWR from 'swr';
 
 const customRepresentation =
